noData handling
when input data is GeoTiff, noDataValue is set to np.nan once data is read as array. This allows to have input data with different data type (int, float).

In KMeans clustering the parameter n_init is explicitely set to the default value of 10 from older sklearn versions to avoid warnings. Could also be set to 'auto' instead of 10.